Add the ability to read a per-user config file located at $HOME/.avrduderc.  Entries from .avrduderc take precedence over those from the system wide config file in ${PREFIX}/etc/avrdude.conf.
Track and display the config file name and line number when we print
out the available parts and programmers.  This is useful in case
someone has overridden a definition in their .avrduderc file and is
wondering why the definition in the system wide config file is not
being used.

Remove the default programmer 'stk500' from the distributed config
file.


git-svn-id: svn://svn.sv.gnu.org/avrdude/trunk@222 81a1dc3b-b13d-400b-aceb-764788c761c2